Populate Marker with DialogBox

+++ app/src/main/java/com/funfit/usjr/thesis/funfitv2/utils/Utils.java
@@ -23,7 +23,7 @@ public class Utils {
     }
 
     //weight(kg), time(hour), speed(km/h)
-    public static int getCaloriesBurned(int weight, int time, int speed){
+    public static int getCaloriesBurned(int weight, long time, float speed){
         //millisecond to hour
         time = (int) ((time / (1000*60*60)) % 24);
         return (int)((0.0215 * (speed*3) - 0.1765 * (speed*2) + 0.8710 * (speed) + 1.4577) * weight * time);
